Logout işlemleri tamamlandı,gitignore eklendi

+++ CloudERP/Controllers/tblUserTypesController.cs
@@ -17,12 +17,22 @@ namespace CloudERP.Controllers
         // GET: tblUserTypes
         public ActionResult Index()
         {
+            if (string.IsNullOrEmpty(Convert.ToString(Session["CompanyID"])))
+            {
+                return RedirectToAction("Login");
+
+            }
             return View(db.tblUserTypes.ToList());
         }
 
         // GET: tblUserTypes/Details/5
         public ActionResult Details(int? id)
         {
+            if (string.IsNullOrEmpty(Convert.ToString(Session["CompanyID"])))
+            {
+                return RedirectToAction("Login");
+
+            }
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
@@ -38,6 +48,11 @@ namespace CloudERP.Controllers
         // GET: tblUserTypes/Create
         public ActionResult Create()
         {
+            if (string.IsNullOrEmpty(Convert.ToString(Session["CompanyID"])))
+            {
+                return RedirectToAction("Login");
+
+            }
             return View();
         }
 
@@ -46,8 +61,13 @@ namespace CloudERP.Controllers
         // Daha fazla bilgi için bkz. https://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public ActionResult Create([Bind(Include = "UserTypeID,UserType")] tblUserType tblUserType)
+        public ActionResult Create(tblUserType tblUserType)
         {
+            if (string.IsNullOrEmpty(Convert.ToString(Session["CompanyID"])))
+            {
+                return RedirectToAction("Login");
+
+            }
             if (ModelState.IsValid)
             {
                 db.tblUserTypes.Add(tblUserType);
@@ -61,6 +81,11 @@ namespace CloudERP.Controllers
         // GET: tblUserTypes/Edit/5
         public ActionResult Edit(int? id)
         {
+            if (string.IsNullOrEmpty(Convert.ToString(Session["CompanyID"])))
+            {
+                return RedirectToAction("Login");
+
+            }
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
@@ -78,8 +103,13 @@ namespace CloudERP.Controllers
         // Daha fazla bilgi için bkz. https://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public ActionResult Edit([Bind(Include = "UserTypeID,UserType")] tblUserType tblUserType)
+        public ActionResult Edit( tblUserType tblUserType)
         {
+            if (string.IsNullOrEmpty(Convert.ToString(Session["CompanyID"])))
+            {
+                return RedirectToAction("Login");
+
+            }
             if (ModelState.IsValid)
             {
                 db.Entry(tblUserType).State = EntityState.Modified;
@@ -92,6 +122,11 @@ namespace CloudERP.Controllers
         // GET: tblUserTypes/Delete/5
         public ActionResult Delete(int? id)
         {
+            if (string.IsNullOrEmpty(Convert.ToString(Session["CompanyID"])))
+            {
+                return RedirectToAction("Login");
+
+            }
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
@@ -109,6 +144,11 @@ namespace CloudERP.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult DeleteConfirmed(int id)
         {
+            if (string.IsNullOrEmpty(Convert.ToString(Session["CompanyID"])))
+            {
+                return RedirectToAction("Login");
+
+            }
             tblUserType tblUserType = db.tblUserTypes.Find(id);
             db.tblUserTypes.Remove(tblUserType);
             db.SaveChanges();
